+++ packages/benchmark-js/src/benchmarkBase.ts
@@ -1,9 +1,0 @@
-export abstract class BenchmarkBase {
-    private static staticId: number = 0;
-
-    protected id: number;
-
-    public constructor() {
-        this.id = ++BenchmarkBase.staticId;
-    }
-}