+++ packages/benchmark-js/src/index.ts
@@ -1,1 +1,3 @@
-export * from './benchmark';
+export * from './Benchmark';
+export * from './BenchmarkGroup';
+export { BenchmarkLoggerLevel } from './Logger';